Formats Post Component using Semantic UI
Installs 'react-time' and 'moment' npm packages to format time, which is
used in the footer of the posts as metadata. Can't quite get the author
for the posts to appear all the way to the right. Custom flexbox styling
supposed to do just that, but I can't figure out which Semantic UI props
are preventing that from happening.