Installs semantic-ui-css locally
Prior to this installation, SUI CSS is being imported in the header of
the React HTML file. I believe this to be causing precedence issues,
whereas I cannot get styles with '!important' to override SUI styles.
Hopefully this will be the solution by importing this CSS in the
index.js file immediately above custom CSS.