More Post styling/formatting
Tweaks grouping and size of containers to format Posts better in Home
and My Posts components. Makes them a little wider, buttons in tight,
keeps margin between posts, replaces Segment with Container. Most
notably, all Posts in My Posts are the same width as the EditPostForm
when it is displayed.