API to localhost, imports new font
Changes API back to development after redeploying. Imports a new font to
use in the body of posts and pages with 'Lato', the supplied SUI default
font. Made Navbar a bit bigger by increasing the font-size of the .logo
as well as the anchors. Need to resize that though for responsiveness.